requirements.txt
asgiref==3.8.1
dj-rest-auth==5.1.0
Django==4.2.11
django-rest-auth==0.9.5
djangorestframework==3.15.1
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
